backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1685243452975/work
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1698832743720/work
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1697026299595/work
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1663854154474/work
frozenlist==1.4.1
greenlet==3.0.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.73
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pgvector==0.2.4
phonenumberslite==8.13.25
Pillow==10.1.0
psycopg @ file:///home/conda/feedstock_root/build_artifacts/psycopg-split_1696017067733/work/psycopg
psycopg-c @ file:///home/conda/feedstock_root/build_artifacts/psycopg-split_1696017067733/work/psycopg-c
psycopg2 @ file:///home/conda/feedstock_root/build_artifacts/psycopg2-split_1697059297905/work
pydantic==2.5.2
pydantic_core==2.14.5
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1693930252784/work
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1681817562700/work
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
urllib3==2.1.0
yarl==1.9.4
